+++ group-1/classes/basket.py
@@ -1,0 +1,4 @@
+class Basket:
+        def __init__(self, basket_item, cost):
+            self.basket_item = basket_item
+            self.cost = cost
