+++ group-1/handler.py
@@ -7,75 +7,16 @@ import boto3
 import numpy as np
 from dotenv import load_dotenv
 from itertools import chain
+from basket import Basket
+from transaction import Transaction
+from core.functions import fill_basket_list, fill_transaction_list
 load_dotenv()
 
 def start(event, context):
 
-    # connect to bucket and get bucket key
-    s3 = boto3.client('s3')
-    response = s3.list_objects(Bucket='cafe-transactions')
-    first_step = response['Contents'][-1]
-    key = first_step['Key']
+    basket_list = fill_basket_list()
+    transactions_list = fill_transaction_list()
 
-    bucket = 'cafe-transactions'
-    path = f's3://{bucket}/{key}'
-    # read bucket contents and sort them into variables
-    df = pd.read_csv(path, names=['date', 'location', 'customer_name', 'basket', 'pay_amount', 'payment_method', 'ccn'])
-    dates = df['date']
-    basket = df['basket']
-    location = df['location']
-    total = df['pay_amount']
-    customer = df['customer_name']
-    customers = list(customer)
-    
-    class Transaction:
-        def __init__(self, total, customer_name, date, location):
-            self.total = total
-            self.customer_name = customer_name
-            self.date = date
-            self.location = location
-        def __repr__(self):
-            return f'Customer name is {self.customer_name}.'
-    # fill transaction list
-    transactions_today = []
-    for i in range(1, len(df)):
-        customer = Transaction(total[i], customers[i], dates[i], location[i])
-        transactions_today.append(customer)
-    
-    
-
-    # create comma-separated strings
-    def chainer(s):
-        return list(chain.from_iterable(s.str.split(',')))
-    # determine lengths of splits
-    lens = df['basket'].str.split(',').map(len)
-    # create new dataframe for chain orders
-    new = pd.DataFrame({'customer_name': np.repeat(df['customer_name'], lens),
-                        'date': np.repeat(df['date'], lens),
-                        'basket': chainer(df['basket'])})
-
-    basket_price = []
-    basket_items = []
-    # getting the price of the basket item and the basket item
-    for i in new['basket']:
-        price = float(i[-5:])
-        basket_item = i[:-5]
-        basket_item_test = basket_item.strip(' -')
-        basket_price.append(price)
-        basket_items.append(basket_item_test)
-    
-
-    class Basket:
-        def __init__(self, basket_item, cost):
-            self.basket_item = basket_item
-            self.cost = cost
-    basket_sep = []
-    # fill our basket list with the basket objects
-    for i in range(0, len(new['basket'])):
-        purchase = Basket(basket_items[i], basket_price[i])
-        basket_sep.append(purchase)
-    
-    
     host = os.getenv("DB_HOST")
     port = int(os.getenv("DB_PORT"))
     user = os.getenv("DB_USER")
@@ -118,7 +59,7 @@ def start(event, context):
                 transaction.customer_name,
                 transaction.date,
                 transaction.location
-            ) for transaction in transactions_today])
+            ) for transaction in transactions_list])
             conn.commit()    
     
     except Exception as ERROR:
@@ -133,7 +74,7 @@ def start(event, context):
             """, [(
                 transaction.basket_item,
                 transaction.cost,
-            ) for transaction in basket_sep])
+            ) for transaction in basket_list])
             conn.commit()  
             conn.close()
